refactor: 장바구니 구매 체크박스 선택 로직 개선
