+++ src/hooks/useCart.ts
@@ -1,15 +1,10 @@
 import {useRecoilState, useRecoilValue} from 'recoil';
 import {allCartCheckedSelector, cartState} from '../recoil/cartAtoms';
-import {initCartListCheckbox, updateCartCheckbox} from '../domain/cart';
+import {initCartListCheckbox} from '../domain/cart';
 
 function useCart() {
   const [cartList, setCartList] = useRecoilState(cartState);
   const isAllCartItemChecked = useRecoilValue(allCartCheckedSelector);
-
-  const switchCheckbox = (id: number) => {
-    const changedCartList = updateCartCheckbox([...cartList], id);
-    setCartList(changedCartList);
-  };
 
   const switchAllCheckboxes = () => {
     const changedCartList = initCartListCheckbox(cartList, !isAllCartItemChecked);
@@ -17,7 +12,6 @@ function useCart() {
   };
 
   return {
-    switchCheckbox,
     switchAllCheckboxes
   };
 }
